Solution to Perform Bitwise OR

+++ Perform Bitwise OR@@67882a731b4392d8fc1577e2/code.c
@@ -3,6 +3,6 @@
 int main(){
     int a,b;
     scnaf("%d %d",&a,&b);
-    printf("%d",a|b);
+    printf("%d", (int)a|b);
     return 0;
 }
